pyproject.toml
[tool.poetry]
name = "async-python-sprint-5"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "^0.27.0"
fastapi-pagination = "^0.11.0"
orjson = "^3.8.3"
pydantic = "^1.10.2"
uvicorn = "^0.20.0"
ormar = "^0.12.0"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.5"
fastapi-users = "^10.2.1"
python-multipart = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
